+++ Modules/_psyco/psyfunc.c
@@ -790,7 +790,7 @@ static PyObject* psycofunction_call(PsycoFunctionObject* self,
 		}
 
 		if (kw != NULL && PyDict_Check(kw)) {
-			int pos, i;
+			Py_ssize_t pos, i;
 			nk = PyDict_Size(kw);
 			k = PyMem_NEW(PyObject *, 2*nk);
 			if (k == NULL) {
